Rename modules/dashboard/pages/dashboard to dashboard-page
