+++ src/app/app-routing.module.ts
@@ -7,7 +7,7 @@ import { AuthGuard } from './core/guards/auth.guard';
 import { GuestGuard } from './core/guards/guest.guard';
 import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
 import { MainLayoutComponent } from './shared/layouts/main-layout/main-layout.component';
-import { DashboardPageComponent } from './modules/dashboard/pages/dashboard/dashboard-page.component';
+import { DashboardPageComponent } from './modules/dashboard/pages/dashboard-page/dashboard-page.component';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
 
 const ROUTES: Routes = [
