+++ src/app/modules/dashboard/dashboard.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DashboardPageComponent } from './pages/dashboard/dashboard-page.component';
+import { DashboardPageComponent } from './pages/dashboard-page/dashboard-page.component';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
